Front-end 18
Details page is started and Mobile view is complete. Home page section components are setup.
