+++ frontend/src/components/user/slider/Slick_slider.jsx
@@ -1,0 +1,94 @@
+import React from "react";
+//slick slider
+import Slider from "react-slick";
+import "slick-carousel/slick/slick.css";
+import "slick-carousel/slick/slick-theme.css";
+
+const Slick_slider = () => {
+  //slick slider settings
+  function SampleNextArrow(props) {
+    const { onClick } = props;
+    return (
+      <div
+        onClick={onClick}
+        className="absolute top-1/2 right-[-20px] z-10 hidden xl:flex items-center justify-center 
+                 w-[40px] h-[40px] rounded-full border-[2px] border-white text-white 
+                 cursor-pointer -translate-y-1/2 bg-black/40 hover:bg-black/70 transition"
+      >
+        <i className="bi bi-chevron-right"></i>
+      </div>
+    );
+  }
+
+  function SamplePrevArrow(props) {
+    const { onClick } = props;
+    return (
+      <div
+        onClick={onClick}
+        className="absolute top-1/2 left-[-20px] z-10 hidden xl:flex items-center justify-center 
+                 w-[40px] h-[40px] rounded-full border-[2px] border-white text-white 
+                 cursor-pointer -translate-y-1/2 bg-black/40 hover:bg-black/70 transition"
+      >
+        <i className="bi bi-chevron-left"></i>
+      </div>
+    );
+  }
+
+  const settings = {
+    dots: false,
+    infinite: true,
+    speed: 1000,
+    slidesToShow: 4,
+    slidesToScroll: 1,
+    autoplay: true,
+    autoplaySpeed: 3000,
+    nextArrow: <SampleNextArrow />,
+    prevArrow: <SamplePrevArrow />,
+    responsive: [
+      { breakpoint: 1260, settings: { slidesToShow: 3, slidesToScroll: 1 } },
+      { breakpoint: 1024, settings: { slidesToShow: 2, slidesToScroll: 1 } },
+      { breakpoint: 600, settings: { slidesToShow: 1, slidesToScroll: 1 } },
+    ],
+  };
+  const movies = [
+    { id: 1, title: "Movie 1", image: "movies/1.jpg" },
+    { id: 2, title: "Movie 2", image: "movies/2.jpg" },
+    { id: 3, title: "Movie 3", image: "movies/3.jpg" },
+    { id: 4, title: "Movie 4", image: "movies/4.jpg" },
+    { id: 5, title: "Movie 5", image: "movies/5.jpg" },
+    { id: 6, title: "Movie 6", image: "movies/6.jpg" },
+    { id: 7, title: "Movie 7", image: "movies/7.jpg" },
+    { id: 8, title: "Movie 8", image: "movies/8.jpg" },
+    { id: 9, title: "Movie 9", image: "movies/9.jpg" },
+  ];
+
+  return (
+    <section className="w-[90%] mx-auto cursor-default xl:mt-[-11vh] xl:z-20 xl:relative ">
+      <h3 className="text-white font-extralight uppercase xl:hidden">
+        Now Showing <i className="bi bi-chevron-right"></i>
+      </h3>
+      <h2 className="hidden text-white font-extralight uppercase xl:block">
+        Now Showing <i className="bi bi-chevron-right"></i>
+      </h2>
+      <div className="w-[85%] mx-auto mt-[30px] xl:w-[90%]">
+        <Slider {...settings}>
+          {/* repeat */}
+          {movies.map((movie) => (
+            <div key={movie.id} className="px-2">
+              <div className="w-auto h-[450px] bg-[#242124] rounded-[10px]">
+                <img
+                  className="w-full h-full object-cover opacity-[0.7] rounded-[10px] hover:opacity-[1] transition duration-300 ease-out"
+                  src={movie.image}
+                  alt="movie"
+                />
+              </div>
+            </div>
+          ))}
+          {/* repeat */}
+        </Slider>
+      </div>
+    </section>
+  );
+};
+
+export default Slick_slider;
